Extrais la ligne de "reponsable sélectionnable" dans un composant
